FIX/FEAT: 타입 및 더미데이터 별도 정리
