+++ src/components/Task/TaskList.tsx
@@ -2,12 +2,6 @@ import { useState } from "react";
 import TaskBox from "./TaskBox";
 import UpdateTaskModal from "../modals/UpdateTaskModal";
 
-interface TaskListProps {
-  name: string;
-  isAll?: boolean;
-  taskInfo: TaskType[];
-}
-
 const TaskList = ({ name, isAll = true, taskInfo }: TaskListProps) => {
   const [selectedTask, setSelectedTask] = useState<{
     id: number;
@@ -17,26 +11,6 @@ const TaskList = ({ name, isAll = true, taskInfo }: TaskListProps) => {
     startDate: string;
     endDate: string;
   } | null>(null);
-
-  // 더미 데이터 (담당자 이름 및 기간 추가)
-  // const dummyTasks = [
-  //   {
-  //     id: 1,
-  //     name: "업무 A",
-  //     memberId: 101,
-  //     memberName: "김철수",
-  //     startDate: "2025-02-11 09:00",
-  //     endDate: "2025-02-11 18:00",
-  //   },
-  //   {
-  //     id: 2,
-  //     name: "업무 B",
-  //     memberId: 202,
-  //     memberName: "이영희",
-  //     startDate: "2025-02-12 10:00",
-  //     endDate: "2025-02-12 17:30",
-  //   },
-  // ];
 
   const openModal = () => {
     // setSelectedTask(dummyTasks[0]); // 임의로 첫 번째 더미 데이터를 선택
