+++ src/components/modals/CreateNotePeriodModal.tsx
@@ -3,15 +3,6 @@ import Button from "../common/Button";
 import DateTimeSelect from "../EditProjectModal/DateTimeSelect";
 import CreateAINoteModal from "./CreateAINoteModal";
 
-interface selectedDateType {
-  year: string;
-  month: string;
-  day: string;
-  hour: string;
-  minute: string;
-  ampm: string;
-}
-
 const CreateNotePeriodModal = ({ onClose }: { onClose: () => void }) => {
   const now = new Date();
   const [selectedStartDate, setSelectedStartDate] = useState<selectedDateType>({
