+++ src/components/modals/EditProjectModal.tsx
@@ -1,3 +1,4 @@
+import { dummy } from "../../dummyData/dummy";
 import Button from "../common/Button";
 import DateTimeSelect from "../EditProjectModal/DateTimeSelect";
 import SelectCategory from "../EditProjectModal/SelectCategory";
@@ -5,153 +6,6 @@ import SelectMember from "../EditProjectModal/SelectMember";
 import WordCloud from "../EditProjectModal/WordCloud";
 import WriteProjectName from "../EditProjectModal/WriteProjectName";
 import { useEffect, useState } from "react";
-
-// (임시) 프로젝트 배열
-const projectData = [
-  {
-    name: "최종프로젝트",
-    startDate: "2025-02-03",
-    endDate: "2025-03-12",
-  },
-  {
-    name: "토이프로젝트",
-    startDate: "2025-02-03",
-    endDate: "2025-03-12",
-  },
-  {
-    name: "사이드프로젝트",
-    startDate: "2025-02-03",
-    endDate: "2025-03-12",
-  },
-  { name: "1차프로젝트", startDate: "2025-02-03", endDate: "2025-03-12" },
-  { name: "2차프로젝트", startDate: "2025-02-03", endDate: "2025-03-12" },
-  { name: "3차프로젝트", startDate: "2025-02-03", endDate: "2025-03-12" },
-  {
-    name: "파이널프로젝트",
-    startDate: "2025-02-03",
-    endDate: "2025-03-12",
-  },
-];
-
-// (임시) 카테고리 데이터
-const categoryData = [
-  {
-    name: "개발",
-    subcategories: [
-      {
-        subname: "사용언어",
-        data: [
-          { text: "C", value: 10 },
-          { text: "C++", value: 20 },
-          { text: "C#", value: 30 },
-          { text: "Java", value: 40 },
-          { text: "JavaScript", value: 50 },
-          { text: "TypeScript", value: 60 },
-          { text: "Python", value: 70 },
-          { text: "Go", value: 80 },
-          { text: "PHP", value: 90 },
-          { text: "Swift", value: 100 },
-          { text: "Kotlin", value: 110 },
-          { text: "기타", value: 120 },
-        ],
-      },
-      {
-        subname: "프레임워크/라이브러리",
-        data: [
-          { text: "Spring", value: 10 },
-          { text: "React", value: 20 },
-          { text: "Vue.js", value: 30 },
-          { text: "Svelte", value: 40 },
-          { text: "Angular", value: 50 },
-          { text: "Flutter", value: 60 },
-          { text: "Next.js", value: 70 },
-          { text: "Nuxt.js", value: 80 },
-          { text: "Unity", value: 90 },
-          { text: "Unreal", value: 100 },
-          { text: "Django", value: 110 },
-          { text: "Flask", value: 120 },
-          { text: "Bootstrap", value: 130 },
-          { text: "Tailwind CSS", value: 140 },
-        ],
-      },
-    ],
-  },
-  { name: "교육" },
-  { name: "금융" },
-  { name: "디자인" },
-  { name: "제조" },
-  { name: "기타" },
-];
-
-// (임시) 팀원 배열
-const membersData = [
-  {
-    id: 1,
-    userName: "홍길동",
-    email: "a@gmail.com",
-    password: "1234",
-    grade: "DISABLE",
-    organization: "데브코스1",
-    profileImage:
-      "https://cdn.pixabay.com/photo/2018/01/15/09/17/woman-3083516_1280.jpg",
-    delete: "ACTIVE",
-  },
-  {
-    id: 2,
-    userName: "홍서범",
-    email: "b@gmail.com",
-    password: "1234",
-    grade: "DISABLE",
-    organization: "데브코스2",
-    profileImage:
-      "https://cdn.pixabay.com/photo/2016/11/21/12/42/beard-1845166_1280.jpg",
-    delete: "ACTIVE",
-  },
-  {
-    id: 3,
-    userName: "홍홍홍",
-    email: "c@gmail.com",
-    password: "1234",
-    grade: "DISABLE",
-    organization: "데브코스3",
-    profileImage:
-      "https://cdn.pixabay.com/photo/2018/01/21/14/16/woman-3096664_1280.jpg",
-    delete: "ACTIVE",
-  },
-];
-// 여기까지 임시 데이터
-
-interface EditProjectModalProps {
-  selectedProjectData?: selectedProjectData;
-  setSelectedProjectData?: React.Dispatch<
-    React.SetStateAction<selectedProjectData>
-  >;
-  setIsEditProjectModal: React.Dispatch<React.SetStateAction<boolean>>;
-  projectMember?: MembersType[];
-  title: string;
-}
-
-interface selectedProjectData {
-  projectName: string;
-  projectStatus: string;
-  createdAt: string;
-  startDate: string;
-  endDate: string;
-  cate: string;
-  subcate1: string[];
-  subcate2: string[];
-}
-
-interface MembersType {
-  id: number;
-  userName: string;
-  email: string;
-  password: string;
-  grade: string;
-  organization: string;
-  profileImage: string;
-  delete: string;
-}
 
 const EditProjectModal = ({
   selectedProjectData,
@@ -222,13 +76,13 @@ const EditProjectModal = ({
   // 선택된 분야의 세부항목 이름
   const selectedSubCate =
     selectedData?.cate &&
-    categoryData
+    dummy.categoryData
       .filter((data) => data.name === selectedData?.cate)[0]
       .subcategories?.map((cate) => cate.subname);
 
   // 선택된 분야의 데이터
   const selectedCateData = selectedData?.cate
-    ? categoryData.filter((data) => data.name === selectedData?.cate)[0]
+    ? dummy.categoryData.filter((data) => data.name === selectedData?.cate)[0]
         .subcategories
     : null;
 
@@ -261,7 +115,7 @@ const EditProjectModal = ({
             <SelectCategory
               selectedData={selectedData}
               setSelectedData={setSelectedData}
-              categoryData={categoryData}
+              categoryData={dummy.categoryData}
             />
           </div>
         )}
@@ -270,7 +124,10 @@ const EditProjectModal = ({
         {pages === 1 && (
           <div className="w-full flex flex-col gap-[20px]">
             {/* 팀원 검색 */}
-            <SelectMember data={membersData} selectedData={projectMember} />
+            <SelectMember
+              data={dummy.membersData}
+              selectedData={projectMember}
+            />
 
             {/* 기간 설정 */}
             <div className="flex flex-col gap-[5px]">
