+++ src/types/alarm.d.ts
@@ -1,0 +1,11 @@
+interface AlarmModalProps {
+  onClose: () => void;
+}
+
+interface AlarmBoxProps {
+  project: string;
+  task?: string;
+  theme: "message" | "newTask" | "newProject" | "endProject";
+  css?: string;
+  onRemove?: () => void;
+}
