+++ src/types/member.d.ts
@@ -1,0 +1,15 @@
+interface MembersType {
+  id: number;
+  userName: string;
+  email: string;
+  password: string;
+  grade: string;
+  organization: string;
+  profileImage: string;
+  delete: string;
+}
+
+interface SelectMembersProps {
+  data: MembersType[];
+  selectedData?: MembersType[];
+}
