+++ src/types/messages.d.ts
@@ -1,0 +1,6 @@
+interface MessageProps {
+  id: number;
+  text: string;
+  sender: string;
+  profile: string;
+}
