+++ src/types/wordCloud.d.ts
@@ -1,0 +1,11 @@
+interface Word {
+  text: string;
+  size: number;
+  x?: number;
+  y?: number;
+  rotate?: number;
+}
+
+interface WordCloudProps {
+  words: { text: string; value: number }[];
+}
